refactor(core): move recovery.ts to recovery/backupKeyRecovery.ts
